# Conflicts:
#	bin/main/com/ceiba/induccion/vehiculos/VehiculoControlador.class

+++ src/main/java/com/ceiba/induccion/utils/factura/Factura.java
@@ -18,9 +18,7 @@ public class Factura {
 	protected int dias = 0;		
 	protected int precioFactura;
 	private Calendar fechaDeIngreso;
-	private Calendar fechaDeSalida;
-	
-	// TODO: delete the reset invocation
+	private Calendar fechaDeSalida;	
 	
 	public void generar(VehiculoModelo vehiculo) {
 		
@@ -58,7 +56,7 @@ public class Factura {
 	
 	private float convertirMiliEnHoras(long milisegundos) {		
 		
-		float equivalencia = 1000 * 60 * 60;  
+		float equivalencia = ((float) 1000) * 60 * 60;  
 		return milisegundos/equivalencia;		
 	}
